requirements.txt
anthropic
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pyodbc
azure-identity
